fastAPI_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
gtts==2.5.1
google-generativeai==0.3.2
pdfplumber==0.10.3
pydub==0.25.1
ffmpeg-python==0.2.0